Add native wind and refactor structure into src folder SVNW-99
