+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mobile-app",
+  "name": "food-order-app",
   "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
@@ -28,13 +28,10 @@
     "expo-symbols": "~0.4.4",
     "expo-system-ui": "~5.0.7",
     "expo-web-browser": "~14.1.6",
-    "lucide-react-native": "^0.509.0",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-native": "0.79.2",
     "react-native-gesture-handler": "~2.24.0",
-    "react-native-reanimated": "~3.17.4",
-    "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.10.0",
     "react-native-web": "~0.20.0",
     "react-native-webview": "13.13.5"
@@ -44,7 +41,13 @@
     "@types/react": "~19.0.10",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~9.2.0",
+    "nativewind": "^4.1.23",
+    "react-native-reanimated": "^3.16.2",
+    "react-native-safe-area-context": "^5.4.0",
+    "tailwindcss": "^3.4.17",
     "typescript": "~5.8.3"
   },
-  "private": true
+  "volta": {
+    "node": "22.10.0"
+  }
 }
