+++ src/app/(tabs)/explore.tsx
