+++ src/components/HapticTab.tsx
