+++ src/components/ui/TabBarBackground.tsx
